+++ MasKit/AppStore/CKSoftwareProduct+SoftwareProduct.swift
@@ -1,0 +1,12 @@
+//
+//  File.swift
+//  MasKit
+//
+//  Created by Ben Chatelain on 12/27/18.
+//  Copyright © 2018 mas-cli. All rights reserved.
+//
+
+import StoreFoundation
+
+// MARK: - SoftwareProduct
+extension CKSoftwareProduct: SoftwareProduct {}