+++ MasKitTests/Commands/SearchSpec.swift
@@ -60,26 +60,3 @@ class SearchSpec: QuickSpec {
         }
     }
 }
-
-/// Nimble predicate for result enum success case, no associated value
-private func beSuccess() -> Predicate<Result<(), MASError>> {
-    return Predicate.define("be <success>") { expression, message in
-        if let actual = try expression.evaluate(),
-                case .success = actual {
-            return PredicateResult(status: .matches, message: message)
-        }
-        return PredicateResult(status: .fail, message: message)
-    }
-}
-
-/// Nimble predicate for result enum failure with associated error
-private func beFailure(test: @escaping (MASError) -> Void = { _ in }) -> Predicate<Result<(), MASError>> {
-    return Predicate.define("be <failure>") { expression, message in
-        if let actual = try expression.evaluate(),
-                case let .failure(error) = actual {
-            test(error)
-            return PredicateResult(status: .matches, message: message)
-        }
-        return PredicateResult(status: .fail, message: message)
-    }
-}